requirements.txt
# Core dependencies
streamlit==1.32.0
sentence-transformers==2.6.0
chromadb==0.4.23
google-generativeai==0.3.2

# Image processing
Pillow==10.2.0
torchvision==0.16.0
torch==2.1.0

# Utilities
tqdm==4.66.2
matplotlib==3.8.2
numpy==1.26.3

# For text processing
transformers==4.36.2
beautifulsoup4==4.12.3

# For debugging and development
ipython==8.22.1